added cart count badge
