+++ apps/client/src/components/navbar/navbar.module.scss
@@ -54,21 +54,23 @@
             gap: 1.5rem;
             align-items: center;
             
-            .cartNumItems {
-                width: 1.5rem;
-                height: 1.5rem;
-                background-color: #ff4a4a;
-                border-radius: 100%;
-                padding: 20px;
-                
-                p {
-                    color: #000000;
-                    font-size: 1rem;
-                    font-weight: 600;
-                    text-align: center;
-                    z-index: 999;
+            .cartIcon {
+                cursor: pointer;
+                position: relative;
+                .cartNumItems {
+                    position: absolute;
+                    top: -10px;
+                    right: -10px;
+                    background-color: #000;
+                    color: #fff;
+                    font-size: 0.5rem;
+                    border-radius: 50%;
+                    height: 1rem;
+                    width: 1rem;
+                    display: flex;
+                    justify-content: center;
+                    align-items: center;
                 }
-                
             }
         }
     }
